Extract CoefficientExponent into its own module
Move the CoefficientExponent class, its rounding helper, the
rounding-mode vocabulary, and formatExponent out of Decimal.mts into a
new src/CoefficientExponent.mts. This restores the file boundary the
class used to have while keeping it an implementation detail: a new
`exports` field in package.json restricts imports to the package's
main entry point, so the module (and any other internal file) cannot be
deep-imported; a new test pins that guarantee via Node package
self-reference. The Stryker build command now transpiles both source
files and mutates both.
