+++ stryker.config.mjs
@@ -3,10 +3,10 @@ export default {
     testRunner: "command",
     commandRunner: {
         command:
-            "./node_modules/.bin/esbuild src/Decimal.mts" +
-            " --format=esm --target=es2020 --outfile=src/Decimal.mjs" +
+            "./node_modules/.bin/esbuild src/Decimal.mts src/CoefficientExponent.mts" +
+            " --format=esm --target=es2020 --outdir=src --out-extension:.js=.mjs" +
             " && ./node_modules/.bin/jest",
     },
-    mutate: ["src/Decimal.mts"],
+    mutate: ["src/*.mts"],
     reporters: ["clear-text", "html", "progress"],
 };
